trade-ping/requirements.txt
flask
grpcio
grpcio-tools
redis
prometheus-client